fixes: correcting functions -> atoi, itoa, memset, strdup, strnstr // wip -> strtrim ans substr

+++ ft_memset.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memset.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gviniciu <gviniciu@student.42.fr>          +#+  +:+       +#+        */
+/*   By: gabriel <gabriel@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/26 14:44:31 by gviniciu          #+#    #+#             */
-/*   Updated: 2026/05/26 19:22:49 by gviniciu         ###   ########.fr       */
+/*   Updated: 2026/06/09 18:14:47 by gabriel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	int		i;
-	char	*ptr;
+	size_t		i;
+	char			*ptr;
 
 	i = 0;
 	ptr = (char *)s;
